+++ src/lua/lua_stack.h
@@ -164,7 +164,7 @@ struct LuaStack
 		va_end(vl);
 	}
 
-	void push_literal(const char* s, size_t len)
+	void push_literal(const char* s, uint32_t len)
 	{
 		lua_pushlstring(L, s, len);
 	}
