+++ src/resource/material_resource.cpp
@@ -202,7 +202,7 @@ namespace material_resource
 
 	void* load(File& file, Allocator& a)
 	{
-		const size_t file_size = file.size();
+		const uint32_t file_size = file.size();
 		void* res = a.allocate(file_size);
 		file.read(res, file_size);
 		return res;
