+++ src/resource/physics_resource.cpp
@@ -339,7 +339,7 @@ namespace physics_resource
 
 	void* load(File& file, Allocator& a)
 	{
-		const size_t file_size = file.size();
+		const uint32_t file_size = file.size();
 		void* res = a.allocate(file_size);
 		file.read(res, file_size);
 		return res;
@@ -713,7 +713,7 @@ namespace physics_config_resource
 
 	void* load(File& file, Allocator& a)
 	{
-		const size_t file_size = file.size();
+		const uint32_t file_size = file.size();
 		void* res = a.allocate(file_size);
 		file.read(res, file_size);
 		return res;
