+++ src/resource/sound_resource.cpp
@@ -70,7 +70,7 @@ namespace sound_resource
 
 	void* load(File& file, Allocator& a)
 	{
-		const size_t file_size = file.size();
+		const uint32_t file_size = file.size();
 		void* res = a.allocate(file_size);
 		file.read(res, file_size);
 		return res;
